74 provide functionality for `list of data frames` (#81)
close #74 

create `ls_explicit_na` to remove na values in list of `data.frame`. The
function uses the `reformat` function to make maintenance easier.
In addition:
- `reformat` is now defined for logicals (i think that was the most
elegant way to do it)

Question:
- Currently, in `reformat` an NA level is created if NA values are
reformatted even if no NA is present in the dataset. I modified it so
that NA level is not created if there are no NA in the original object
to be more consistant with the rest of the reformatting, but we could go
both way. What so you think @clarkliming ?

Example:
before:
```
x = c("a", "b")
rr <- rule("<Missing>" = NA)
reformat(x, rr)
"a" "b"
levels: "a", "b", "<Missing>"
```
now

```
x = c("a", "b")
rr <- rule("<Missing>" = NA)
reformat(x, rr)
"a" "b"
levels: "a", "b"
```

---------

Signed-off-by: b_falquet <64274616+BFalquet@users.noreply.github.com>
Co-authored-by: benoit <benoit.falquet@roche.com>
Co-authored-by: Liming <36079400+clarkliming@users.noreply.github.com>